Use build constraints for `Socket` implementations
